Move cli/util command hierarchy related stuff to cli/cmd.go.
